// src/types/dataTypes.ts
export interface formData {
    key: number
    title: string
    firstName: string
    lastName: string
    fullName: string
    birthDay: string
    nationality: string
    citizenID: string
    gender: string
    passport: string
    callingCode: string
    phoneNumber: string
    phone: string
    salary: string
}